Split off graphics code for units
