+++ emergence_lib/src/graphics/units.rs
@@ -1,0 +1,29 @@
+//! Graphics and animation code for units.
+
+use bevy::prelude::*;
+
+use crate::{
+    asset_management::units::UnitHandles,
+    organisms::units::UnitId,
+    simulation::geometry::{MapGeometry, TilePos},
+};
+
+/// Adds rendering components to every spawned unit
+pub(super) fn populate_units(
+    new_units: Query<(Entity, &TilePos, &UnitId), Added<UnitId>>,
+    mut commands: Commands,
+    unit_handles: Res<UnitHandles>,
+    map_geometry: Res<MapGeometry>,
+) {
+    for (entity, tile_pos, unit_id) in new_units.iter() {
+        let pos = map_geometry.layout.hex_to_world_pos(tile_pos.hex);
+        let terrain_height = *map_geometry.height_index.get(tile_pos).unwrap();
+        let scene_handle = unit_handles.scenes.get(unit_id).unwrap();
+
+        commands.entity(entity).insert(SceneBundle {
+            scene: scene_handle.clone_weak(),
+            transform: Transform::from_xyz(pos.x, terrain_height, pos.y),
+            ..default()
+        });
+    }
+}
